index

+++ src/index.ts
@@ -25,8 +25,8 @@ export interface ITranspilerConfig {
   options: Transpiler.ITranspilerOptions;
 }
 
-const COMPILE_RESULT = "_compile.txt";
-const RUN_RESULT = "foobar.txt";
+const COMPILE_RESULT = "_compile_result.txt";
+const RUN_RESULT = "_run_result.txt";
 
 function run() {
 
@@ -47,9 +47,12 @@ function run() {
 
   fs.writeFileSync(outputDir + "/abap_transpile.json", JSON.stringify(config, null, 2));
 
-  execSync(`rm -f ${outputDir}/*.abap`, {stdio: 'pipe'});
-  execSync(`cp ${inputDir}/* ${outputDir}`, {stdio: 'pipe'});
-  execSync(`cp open-abap/src/unit/*.clas.abap ${outputDir}`, {stdio: 'pipe'});
+  if (inputDir !== outputDir) {
+    execSync(`rm -f ${outputDir}/*.abap`, {stdio: 'pipe'});
+    execSync(`cp ${inputDir}/* ${outputDir}`, {stdio: 'pipe'});
+  }
+  execSync(`mkdir -f ${outputDir}/deps/`, {stdio: 'pipe'});
+  execSync(`cp open-abap/src/unit/*.clas.abap ${outputDir}/deps/`, {stdio: 'pipe'});
   execSync(`cp open-abap/src/classrun/*.intf.abap ${outputDir}`, {stdio: 'pipe'});
 
   const output: IOutput = {
